chore: organise folder structure
