+++ apps/journeys-admin/src/components/Editor/Toolbar/JourneyDetails/index.ts
